+++ app/components/userFeedback/FeedbackFormBox.tsx
@@ -5,7 +5,7 @@ import { useEffect, useRef } from "react";
 import { z } from "zod";
 import Textarea from "~/components/inputs/Textarea";
 import { FeedbackTitle } from "~/components/userFeedback/FeedbackTitle";
-import { useJsAvailable } from "~/services/useJsAvailabe";
+import { useJsAvailable } from "~/services/useJsAvailable";
 import { TEXTAREA_CHAR_LIMIT } from "~/services/validation/inputlimits";
 import { useFeedbackTranslations } from "./feedbackTranslations";
 import { FeedbackType } from "./FeedbackType";
