+++ app/services/useJsAvailable.ts
